Adjust phpcs-scan to submit reviews to Pull Requests
Previously, phpcs-scan submitted inline comments on individual commits.
However, as that did not work well with the GitHub user-interface, this has
been changed to submit comments to Pull Requests.

The invocation of phpcs-scan has not changed due to this change. The changes
make sure that it finds all relevant Pull-Requests (i.e. that involve the
commit specified on invocation), then scans every file that is part of the
commit, finally then submits a whole Pull Request review to GitHub containing
every issue found.

This is not commplete, there are some issues still.