Properly sleep inbetween GitHub requests, retry submitting review if needed
A central function was aded to manage sleeping in between GitHub requests.
GitHub requries calls to have a certain waiting period inbetween them, and
we honor that with this function. It only sleeps if it is really needed.

If submitting a review fails, retry after a certain period. Try to detect
if GitHub wants a certain time to pass before retrying