Integrate PHP linting fully, abstract into functions, better error-detection
Integrating PHP linting, using generic structures to contain and submit PHP
linting issues to GitHub -- these are the same as the PHPCS-scanning process
uses.

Also, PHP linting enhanced, breaking the whole process into separate functions.

Error-detection was enhanced when fetching commits which are not found, in
those cases, execution is halted with an error-message