Move check for empty Pull-Request array from phpcs-scan.php to vip-go-ci.php
This check should be performed before doing any scanning, to avoid
doing any work that cannot be acted up on. I.e., if there are no
Pull-Requests that refer to the commit, we cannot comment on the PR
in case of any issues and thus scanning would be without use.