Stats-init to function, check validity of git-repo centralized, latest-commit check
Statistics initialization moved to function to de-clutter the main function.

Instead of doing a primitive check for the validity of the git-repository,
do a more comprehensive one right from the beginning.

Check if the commit we are working with is the latest one, and if not, make
sure not to lint code.