Revert from using GitHub API as fallback
Stop using GitHub API as fallback, if the local repository is unavailable.
This is because maintaining the fallback will add more work as more features
are added; having to maintain the fallback and the local-git repository
support at the same time is not worth the work.

As a result, make the local-git repository mandatory.