Fetching PR-comments changed; fetch all of them and cache for later usage
Instead of repeatedly fetching Pull-Request comments, simply fetch all that
are made after a certain timestamp, and cache these. Then filter out the
ones that are irrelevant and return the result. The cached version can be
used in later calls to the function if the timestamp matches, then the
filtering is done afterwards.