Splitting auto-approval into three modules
One module will oversee the actual approval, based up on if all
files in the PR have been approved by the other modules. These
other modules will look through PR-altered files, and determine
if each file can be auto-approved, and if so, adds it to a list
that the overseeing module looks at.