Whitespacing, adding support for 'modular auto-approvals'.
Modular auto-approvals, essentially, means that auto-approvals
happen within modules that individually decide which files can
be approved. Then, when all modules have finished their work,
a process goes through all the files altered by the PR, and
figures out to do with the whole PR, based on this. If all the
files are approved, the whole PR gets approved, but if only
a subset of files is approved, then a message is attached
to the file (in case of PHP and JS files).

This functionality is not complete, but well underway.