Stop relying on 'all-files' array, and rely only on PR-specific array.
This fixes a bug where two or more Pull-Requests (PRs), using different bases,
might have different files changed, and due to the bug, one of them
might not be its files scanned. This was because we relied on the 'all-files'
array to check if the file had been considered for scanning, and because
the file already was in the first PR, it would not be added to the second one.